SIG-141 rename get_signals to get_auth_token (that is what it does)
